Ensure that the hooks order controller calls the generator
