+++ app/controllers/spree/retail/shopify/hooks/orders_controller.rb
@@ -4,7 +4,14 @@ module Spree
       module Hooks
         class OrdersController < HooksController
           def create
+            shopify_order = ShopifyAPI::Order.find(shopify_order_id)
+            Spree::Retail::Shopify::GeneratePosOrder.new(shopify_order).process
+
             head :ok
+          end
+
+          def shopify_order_id
+            json_body['order']['id']
           end
         end
       end
